requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
decorator==4.4.2
Django==3.1.5
django-filter==2.4.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
# Editable install with no version control (MyCart==1.0)
-e c:\users\paridhi\desktop\arya\mycartapp
mysqlclient==2.0.3
numpy==1.19.5
pandas==1.2.1
psycopg2==2.8.6
pyfiglet==0.8.post1
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.10
validators==0.18.2
